requirements.txt
asgiref==3.6.0
black==22.12.0
click==8.1.3
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.0
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
